Add files via upload
The objective of autonomous driving is to transform transportation by using cutting-edge 
technologies to increase efficiency and safety. This work uses a dataset of driving view images 
mapped to corresponding steering angles that was created by a simulator to investigate the use of 
convolutional neural networks (CNNs) for autonomous driving. Predicting steering angles 
accurately is essential to preserving lane discipline and vehicle safety. To increase model 
generalization and robustness against changing driving conditions, the dataset received substantial 
preprocessing and augmentation, including normalization, scaling, and random alterations. 
Three CNN models—MobileNet-V2, an ensemble approach (Ensemble-M), and Nvidia's 
proposed CNN—were put into practice and compared. To capitalize on their respective strengths, 
Ensemble-M coupled Nvidia's end-to-end learning model with MobileNet-V2, which is renowned 
for striking a balance between computational efficiency and performance. Mean Squared Error 
(MSE) was used to evaluate the models in order to determine their computational efficiency and 
prediction accuracy for real-time deployment. The MSEs for the models from Nvidia and 
MobileNet-V2 were 0.3002, 2.2894, and 1.0014, respectively. 
The results showed that because there were more visual cues on roads with dense features, 
MobileNet-V2 performed better there, whereas Nvidia's model performed better on roads with 
sparse features. Even though Ensemble-M used the most processing power, it produced the lowest 
MSE score, highlighting the advantages of ensemble learning. This research contributes to the 
development of more dependable autonomous driving systems by highlighting the possibility of 
mixing various CNN designs to enhance performance across a range of driving scenarios. In the 
future, studies may investigate the integration of other sensory inputs, including LiDAR and 
RADAR, to improve the performance and resilience of the model and pave the way for safer and 
more effective autonomous cars.